+++ src/common/enums/index.ts
@@ -1,3 +1,2 @@
 export * from './request-method.enum';
-export * from './transport.enum';
 export * from './http-status.enum';