+++ src/common/index.ts
@@ -1,3 +1,10 @@
+/*
+ * Nest @common
+ * Copyright(c) 2017-... Kamil Mysliwiec
+ * www.nestjs.com || www.kamilmysliwiec.com
+ * MIT Licensed
+ */
+
 export * from './utils';
 export * from './enums';
 export {
@@ -5,7 +12,10 @@ export {
     INestApplication,
     INestMicroservice,
     MiddlewareConfigProxy,
+    MiddlewareConfiguration,
+    NestMiddleware,
     MiddlewaresConsumer,
     OnModuleInit,
     ExceptionFilter,
 } from './interfaces';
+export * from './services/logger.service';