+++ src/common/interfaces/nest-module.interface.ts
@@ -1,6 +1,0 @@
-import { MiddlewareBuilder } from '../../core/middlewares/builder';
-import { MiddlewaresConsumer } from './middlewares-consumer.interface';
-
-export interface NestModule {
-    configure?: (consumer: MiddlewaresConsumer) => MiddlewaresConsumer | void;
-}