+++ src/core/errors/exception-handler.ts
@@ -1,5 +1,5 @@
 import { RuntimeException } from './exceptions/runtime.exception';
-import { Logger } from '../common/services/logger.service';
+import { Logger } from '@nestjs/common/services/logger.service';
 
 export class ExceptionHandler {
     private readonly logger = new Logger(ExceptionHandler.name);