+++ src/core/injector/instance-loader.ts
@@ -1,12 +1,12 @@
+import iterate from 'iterare';
 import { NestContainer } from './container';
 import { Injector } from './injector';
-import { Injectable } from '../../common/interfaces/injectable.interface';
-import { Controller } from '../../common/interfaces/controller.interface';
+import { Injectable } from '@nestjs/common/interfaces/injectable.interface';
+import { Controller } from '@nestjs/common/interfaces/controllers/controller.interface';
 import { Module } from './module';
-import { Logger } from '../../common/services/logger.service';
+import { Logger, OnModuleInit } from '@nestjs/common';
 import { ModuleInitMessage } from '../helpers/messages';
-import { isUndefined, isNil } from '../../common/utils/shared.utils';
-import { OnModuleInit } from '../../common/interfaces/index';
+import { isUndefined, isNil } from '@nestjs/common/utils/shared.utils';
 
 export class InstanceLoader {
     private injector = new Injector();
@@ -29,11 +29,12 @@ export class InstanceLoader {
     }
 
     private createInstances(modules: Map<string, Module>) {
-        modules.forEach((module, name) => {
+        modules.forEach((module) => {
             this.createInstancesOfComponents(module);
             this.createInstancesOfRoutes(module);
             this.callModuleInitHook(module);
 
+            const { name } = module.metatype;
             this.logger.log(ModuleInitMessage(name));
         });
     }
@@ -64,7 +65,7 @@ export class InstanceLoader {
 
     private callModuleInitHook(module: Module) {
         const components = [...module.routes, ...module.components];
-        components.map(([key, {instance}]) => instance)
+        iterate(components).map(([key, {instance}]) => instance)
                 .filter((instance) => !isNil(instance))
                 .filter(this.hasOnModuleInitHook)
                 .forEach((instance) => (instance as OnModuleInit).onModuleInit());