+++ src/core/middlewares/routes-mapper.ts
@@ -1,9 +1,9 @@
 import 'reflect-metadata';
 import { ExpressRouterExplorer } from '../router/router-explorer';
-import { UnknownRequestMappingException } from '../../errors/exceptions/unknown-request-mapping.exception';
-import { RequestMethod } from '../../common/enums/request-method.enum';
-import { isUndefined, validatePath } from '../../common/utils/shared.utils';
-import { PATH_METADATA } from '../../common/constants';
+import { UnknownRequestMappingException } from '../errors/exceptions/unknown-request-mapping.exception';
+import { RequestMethod } from '@nestjs/common/enums/request-method.enum';
+import { isUndefined, validatePath } from '@nestjs/common/utils/shared.utils';
+import { PATH_METADATA } from '@nestjs/common/constants';
 import { MetadataScanner } from '../metadata-scanner';
 
 export class RoutesMapper {