+++ src/core/nest-microservice.ts
@@ -1,12 +1,12 @@
-import { NestContainer } from './core/injector/container';
-import { MicroservicesModule } from './microservices/microservices-module';
-import { messages } from './core/constants';
-import { Logger } from './common/services/logger.service';
-import { Server } from './microservices/server/server';
-import { MicroserviceConfiguration } from './microservices/interfaces/microservice-configuration.interface';
-import { ServerFactory } from './microservices/server/server-factory';
-import { Transport } from './common/enums/transport.enum';
-import { INestMicroservice } from './common/interfaces';
+import { NestContainer } from './injector/container';
+import { MicroservicesModule } from '@nestjs/microservices/microservices-module';
+import { messages } from './constants';
+import { Logger } from '@nestjs/common/services/logger.service';
+import { Server } from '@nestjs/microservices/server/server';
+import { MicroserviceConfiguration } from '@nestjs/microservices/interfaces/microservice-configuration.interface';
+import { ServerFactory } from '@nestjs/microservices/server/server-factory';
+import { Transport } from '@nestjs/common/enums/transport.enum';
+import { INestMicroservice } from '@nestjs/common';
 
 export class NestMicroservice implements INestMicroservice {
     private readonly logger = new Logger(NestMicroservice.name);