+++ src/errors/exceptions/invalid-module-config.exception.ts
@@ -1,8 +1,0 @@
-import { RuntimeException } from './runtime.exception';
-import { InvalidModuleConfigMessage } from '../messages';
-
-export class InvalidModuleConfigException extends RuntimeException {
-    constructor(property: string) {
-        super(InvalidModuleConfigMessage(property));
-    }
-}