+++ src/microservices/client/client-proxy-factory.ts
@@ -1,7 +1,7 @@
 import { ClientTCP } from './client-tcp';
 import { ClientRedis } from './client-redis';
 import { ClientMetadata } from '../interfaces/client-metadata.interface';
-import { Transport } from '../../common/enums/transport.enum';
+import { Transport } from '@nestjs/common/enums/transport.enum';
 import { ClientProxy } from './client-proxy';
 
 export class ClientProxyFactory {