+++ src/websockets/index.ts
@@ -1,2 +1,9 @@
+/*
+ * Nest @websockets
+ * Copyright(c) 2017-... Kamil Mysliwiec
+ * www.nestjs.com || www.kamilmysliwiec.com
+ * MIT Licensed
+ */
+
 export * from './interfaces';
 export * from './utils';