+++ src/websockets/middlewares-injector.ts
@@ -1,12 +1,13 @@
 import 'reflect-metadata';
-import { NestContainer, InstanceWrapper } from '../core/injector/container';
+import iterate from 'iterare';
+import { NestContainer, InstanceWrapper } from '@nestjs/core/injector/container';
 import { NestGateway } from './index';
 import { GATEWAY_MIDDLEWARES } from './constants';
-import { UnknownModuleException } from '../errors/exceptions/unknown-module.exception';
-import { Injectable } from '../common/interfaces/injectable.interface';
-import { RuntimeException } from '../errors/exceptions/runtime.exception';
+import { UnknownModuleException } from '@nestjs/core/errors/exceptions/unknown-module.exception';
+import { Injectable } from '@nestjs/common/interfaces/injectable.interface';
+import { RuntimeException } from '@nestjs/core/errors/exceptions/runtime.exception';
 import { GatewayMiddleware } from './interfaces/gateway-middleware.interface';
-import { isUndefined, isFunction, isNil } from '../common/utils/shared.utils';
+import { isUndefined, isFunction, isNil } from '@nestjs/common/utils/shared.utils';
 
 export class MiddlewaresInjector {
     constructor(private container: NestContainer) {}
@@ -27,9 +28,9 @@ export class MiddlewaresInjector {
     }
 
     public applyMiddlewares(server, components: Map<string, InstanceWrapper<Injectable>>, tokens: any[]) {
-        tokens.map(token => this.bindMiddleware(token.name, components))
+        iterate(tokens).map(token => this.bindMiddleware(token.name, components))
             .filter(middleware => !isNil(middleware))
-            .map(middleware => server.use(middleware));
+            .forEach(middleware => server.use(middleware));
     }
 
     public bindMiddleware(token: string, components: Map<string, InstanceWrapper<Injectable>>) {